WIP: EditProfile component.
